Fix for url changes not correctly updating overlayed observation tiles on Android
